Add Note.java, NoteList.java, modify others
Some items may be notes, may not be tasks.
These are simply stuff we want to take down for reference.
Our bot cannot accommodate for these items.

In particular, these items do not need to be marked as done
or undone. They have no related dates or times.
They are just textual information.

Let's
* Add Note.java to represent a Note instance
* Add NoteList.java to represent a list of notes
* Add logic in Parser to add a note, delete a note,
and list all notes
* Update Storage.java, Phil.java, PhilTextUI.java to account
for additional storage of notes in a separate file
* Add corresponding enums in CommandType
* Update IllegalCommandException and IllegalArgumentException
for the possible exceptions from the related actions
* Add JavaDoc comments for the new classes and methods