TaskList.java: Update filteredTasksToString
Users may want to use the find function to get the
real index of the corresponding tasks with a search term
so they can mark it as done or undone.
The current filteredTasksToString uses a temporary count
as the index displayed from find command instead.

Let's
* Edit filteredTasksToString to show the actual
index of tasks matching the search term